Add files via upload
The sudoku game known to everyone.
Pressing "Start Game" will generate a Sudoku board with 40 missing boxes. From the right, set the value you want to put on the board, if the value is not correct, the application will indicate the line or column where it is not correct. With the right click you will be able to remove the set value. Pressing the "Show Solution" button will reveal the game solution.